faq_api/requirements.txt
Django
djangorestframework
redis
googletrans==4.0.0-rc1
django-ckeditor